requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
boto3==1.12.33
botocore==1.15.33
bson==0.5.8
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
cycler==0.10.0
daphne==2.5.0
dataclasses==0.6
dj-database-url==0.5.0
Django==2.2.13
django-heroku==0.3.1
django-storages==1.9.1
djangorestframework==3.11.0
djongo==1.3.2
docutils==0.15.2
elasticsearch==7.8.0
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jmespath==0.9.5
kiwisolver==1.1.0
matplotlib==3.1.3
msgpack==1.0.0
mysqlclient==1.4.6
numpy==1.18.1
pandas==1.0.1
passlib==1.7.2
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
PyHamcrest==2.0.2
PyJWT==1.7.1
pymongo==3.10.1
pyOpenSSL==19.1.0
pyotp==2.3.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
razorpay==1.2.0
requests==2.23.0
s3transfer==0.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.2.4
twilio==6.38.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8
whitenoise==5.0.1
zope.interface==5.1.0
